aws-cost-report/requirements.txt
# Core AWS interaction
boto3==1.34.69
botocore==1.34.69

# Data manipulation
pandas==2.2.1
numpy==1.26.4
python-dateutil==2.8.2

# Visualization
plotly==5.19.0
kaleido==0.2.1          # For static image export (optional)

# Excel output
openpyxl==3.1.2
xlsxwriter==3.1.9

# Web dashboard components
jinja2==3.1.3
markupsafe==2.1.5

# Optional utilities
pytz==2024.1            # Timezone support
tqdm==4.66.2            # Progress bars (optional)